+++ Code/Assets/Scripts/Rule/BattleRule_Defend.cs
@@ -115,7 +115,7 @@ public class Battle_Defend : ABattleRule
 
     private void BuildReward()
     {
-        List<KeyValuePair<int, DropConfig>> dropList = DropLimitHelper.Build((int)DropLimitType.Defend, 1);
+        List<KeyValuePair<double, DropConfig>> dropList = DropLimitHelper.Build((int)DropLimitType.Defend, 1);
 
         List<Item> items = DropHelper.BuildDropItem(dropList, 1);
